Rename python -> src
