+++ setup.cfg
@@ -11,5 +11,5 @@ exclude =
   .eggs
 
 [tool:pytest]
-addopts = --cov=./python/lsst/cm/tools/core --cov=./python/lsst/cm/tools/db --cov=./python/lsst/cm/tools/examples --cov-report=html
+addopts = --cov=./src/lsst/cm/tools/core --cov=./src/lsst/cm/tools/db --cov=./src/lsst/cm/tools/examples --cov-report=html
 flake8-ignore = E133 E203 E226 E228 N802 N803 N806 N812 N813 N815 N816 W503
