+++ setup.py
@@ -6,12 +6,12 @@ setup(
     author_email="echarles@slac.stanford.edu, fritzm@SLAC.Stanford.EDU",
     url="https://github.com/lsst-dm/cm_tools",
     packages=["lsst.cm.tools"],
-    package_dir={"": "python"},
+    package_dir={"": "src"},
     description="Campaign Managment Tools",
     setup_requires=["setuptools_scm"],
     long_description=open("README.rst").read(),
     package_data={"": ["README.rst", "LICENSE"]},
-    use_scm_version={"write_to": "python/lsst/cm/tools/_version.py"},
+    use_scm_version={"write_to": "src/lsst/cm/tools/_version.py"},
     include_package_data=True,
     classifiers=[
         "Development Status :: 4 - Beta",
