+++ src/lsst/cm/tools/db/script.py
@@ -1,22 +1,25 @@
+from sqlalchemy import Boolean, Column, Enum, ForeignKey, Integer, String
+from sqlalchemy.orm import composite, relationship
+
 from lsst.cm.tools.core.db_interface import ScriptBase
 from lsst.cm.tools.core.dbid import DbId
-from lsst.cm.tools.core.utils import LevelEnum, ScriptMethod, StatusEnum
+from lsst.cm.tools.core.utils import LevelEnum, ScriptMethod, ScriptType, StatusEnum
 from lsst.cm.tools.db import common
 from lsst.cm.tools.db.campaign import Campaign
 from lsst.cm.tools.db.group import Group
 from lsst.cm.tools.db.step import Step
 from lsst.cm.tools.db.workflow import Workflow
-from sqlalchemy import Boolean, Column, DateTime, Enum, Float, ForeignKey, Integer, String
-from sqlalchemy.orm import composite, relationship
 
 
-class Job(common.Base, common.SQLScriptMixin, ScriptBase):
-    """Database table to manage processing `Job`
+class Script(common.Base, common.SQLScriptMixin, ScriptBase):
+    """Database table to manage processing `Script`
 
-    A `Job` launchs job on a batch system
+    A `Script` is a small shell script that helps with
+    collection and file management, or does stuff that
+    we want to run independently.
     """
 
-    __tablename__ = "job"
+    __tablename__ = "script"
 
     id = Column(Integer, primary_key=True)  # Unique script ID
     c_id = Column(Integer, ForeignKey(Campaign.id))
@@ -30,35 +33,25 @@ class Job(common.Base, common.SQLScriptMixin, ScriptBase):
     script_url = Column(String)  # Url for script
     stamp_url = Column(String)  # Url for a status 'stamp' file
     log_url = Column(String)  # Url for log
-    config_url = Column(String)  # Url for script configuration
     coll_out = Column(String)  # Output collection
     checker = Column(String)  # Checker class
     rollback = Column(String)  # Rollback class
     status = Column(Enum(StatusEnum))  # Status flag
     superseded = Column(Boolean)  # Has this been superseded
+    script_type = Column(Enum(ScriptType))  # What sort of thing the script does
     script_method = Column(Enum(ScriptMethod))  # How the script is invoked
-    n_tasks_all = Column(Integer, default=0)  # Number of associated tasks
-    n_tasks_done = Column(Integer, default=0)  # Number of finished tasks
-    n_tasks_failed = Column(Integer, default=0)  # Number of failed tasks
-    n_clusters_all = Column(Integer, default=0)  # Number of associated clusters
-    n_clusters_done = Column(Integer, default=0)  # Number of finished clusters
-    n_clusters_failed = Column(Integer, default=0)  # Number of failed clusters
-    job_start = Column(DateTime)  # Workflow start time
-    job_end = Column(DateTime)  # Workflow end time
-    job_cputime = Column(Float)
-
     level = Column(Enum(LevelEnum))
     db_id: DbId = composite(DbId, c_id=c_id, s_id=s_id, g_id=g_id, w_id=w_id)
-    c_: Campaign = relationship("Campaign", back_populates="jobs_")
-    s_: Step = relationship("Step", back_populates="jobs_")
-    g_: Group = relationship("Group", back_populates="jobs_")
-    w_: Workflow = relationship("Workflow", back_populates="jobs_")
+    c_: Campaign = relationship("Campaign", back_populates="scripts_")
+    s_: Step = relationship("Step", back_populates="scripts_")
+    g_: Group = relationship("Group", back_populates="scripts_")
+    w_: Workflow = relationship("Workflow", back_populates="scripts_")
 
     def __repr__(self) -> str:
         if self.superseded:
             supersede_string = "SUPERSEDED"
         else:
             supersede_string = ""
-        string_out = f"BatchJob {self.id}: {self.db_id} {self.name} {self.handler} {self.status.name}"
-        string_out += f" {supersede_string}"
-        return string_out
+        return (
+            f"Script {self.id}: {self.db_id} {self.name} {self.handler} {self.status.name} {supersede_string}"
+        )
