+++ src/lsst/cm/tools/db/step.py
@@ -1,31 +1,31 @@
 from typing import Any, Iterable, TextIO
 
+from sqlalchemy import Boolean, Column, Enum, ForeignKey, Integer, String
+from sqlalchemy.ext.hybrid import hybrid_property
+from sqlalchemy.orm import composite, relationship
+
 from lsst.cm.tools.core.dbid import DbId
 from lsst.cm.tools.core.utils import InputType, LevelEnum, OutputType, StatusEnum
 from lsst.cm.tools.db import common
 from lsst.cm.tools.db.campaign import Campaign
 from lsst.cm.tools.db.production import Production
-from lsst.cm.tools.db.step import Step
-from sqlalchemy import Boolean, Column, Enum, ForeignKey, Integer, String
-from sqlalchemy.ext.hybrid import hybrid_property
-from sqlalchemy.orm import composite, relationship
 
 
-class Group(common.Base, common.CMTable):
-    """Database table to manage processing `Group`
+class Step(common.Base, common.CMTable):
+    """Database table to manage processing `Step`
 
-    A `Group` can be processed in a single `Workflow`,
-    but we also need to account for possible failures.
+    A `Step` consists of several processing `Group` which
+    are run in parallel
+
     """
 
-    __tablename__ = "group"
+    __tablename__ = "step"
 
-    level = LevelEnum.group
-    id = Column(Integer, primary_key=True)  # Unique Group ID
+    level = LevelEnum.step
+    id = Column(Integer, primary_key=True)  # Unique Step ID
     p_id = Column(Integer, ForeignKey(Production.id))
     c_id = Column(Integer, ForeignKey(Campaign.id))
-    s_id = Column(Integer, ForeignKey(Step.id))
-    name = Column(String)  # Group name
+    name = Column(String)  # Step name
     fullname = Column(String, unique=True)  # Unique name
     handler = Column(String)  # Handler class
     config_yaml = Column(String)  # Configuration file
@@ -38,16 +38,17 @@ class Group(common.Base, common.CMTable):
     output_type = Column(Enum(OutputType))  # How to manage output data
     status = Column(Enum(StatusEnum))  # Status flag
     superseded = Column(Boolean)  # Has this been superseded
-    db_id: DbId = composite(DbId, p_id, c_id, s_id, id)
+    previous_step_id = Column(Integer)
+    db_id: DbId = composite(DbId, p_id, c_id, id)
     p_: Production = relationship("Production", foreign_keys=[p_id])
-    c_: Campaign = relationship("Campaign", back_populates="g_")
-    s_: Step = relationship("Step", back_populates="g_")
-    w_: Iterable = relationship("Workflow", back_populates="g_")
-    scripts_: Iterable = relationship("Script", back_populates="g_")
-    jobs_: Iterable = relationship("Job", back_populates="g_")
-    depend_: Iterable = relationship("Dependency", back_populates="g_")
+    c_: Campaign = relationship("Campaign", back_populates="s_")
+    g_: Iterable = relationship("Group", back_populates="s_")
+    w_: Iterable = relationship("Workflow", back_populates="s_")
+    scripts_: Iterable = relationship("Script", back_populates="s_")
+    jobs_: Iterable = relationship("Job", back_populates="s_")
+    depend_: Iterable = relationship("Dependency", back_populates="s_")
 
-    match_keys = [p_id, c_id, s_id, id]
+    match_keys = [p_id, c_id, id]
 
     @hybrid_property
     def butler_repo(self) -> Any:
@@ -60,30 +61,35 @@ class Group(common.Base, common.CMTable):
         return self.c_.prod_base_url
 
     @hybrid_property
+    def root_coll(self) -> Any:
+        """Direct access to the root of collection names, for convinience"""
+        return self.c_.root_coll
+
+    @hybrid_property
     def parent_id(self) -> Any:
-        """Maps s_id to parent_id for consistency"""
-        return self.s_id
+        """Maps c_id to parent_id for consistency"""
+        return self.c_id
 
     def __repr__(self) -> str:
         if self.superseded:
             supersede_string = "SUPERSEDED"
         else:
             supersede_string = ""
-        return f"Group {self.fullname} {self.db_id}: {self.handler} {self.status.name} {supersede_string}"
+        return f"Step {self.fullname} {self.db_id}: {self.handler} {self.status.name} {supersede_string}"
 
     def print_tree(self, stream: TextIO) -> None:
         """Print entry in tree-like format"""
-        stream.write(f"    {self}\n")
+        stream.write(f"  {self}\n")
         for script in self.scripts_:
-            stream.write(f"      -{script}\n")
-        for workflow in self.w_:
-            stream.write(f"      {workflow}")
+            stream.write(f"    -{script}\n")
+        for group in self.g_:
+            group.print_tree(stream)
 
     def children(self) -> Iterable:
-        """Maps self.w_ to self.children() for consistency"""
-        for workflow in self.w_:
-            yield workflow
+        """Maps self.g_ to self.children() for consistency"""
+        for group in self.g_:
+            yield group
 
     def sub_iterators(self) -> list[Iterable]:
         """Iterators over sub-entries, used for recursion"""
-        return [self.w_]
+        return [self.w_, self.g_]
