+++ src/lsst/cm/tools/version.py
